Calculate Daily Returns

Calculate day-over-day historical returns for a list of assets, taking into account splits and dividends.

calculate_daily_returns(assets, date_range_xts = NULL,
  buy_at = c("close", "open", "high", "low", "price"),
  sell_at = c("close", "open", "high", "low", "price"),
  include_volume = TRUE, include_buy_sell = FALSE,
  returns_method = c("ln", "log2", "log10", "pct_diff", "multiple"))

Arguments

assets
a list of sub-lists named and corresponding to a given asset, for example, "AAPL". Each sublist contains split, dividend, and aggregated OHLCV data as xts objects. An example of an object that may be passed as asset_data is sample_historical_data.

date_range_xts
A string that specifies the date range for which calculate_daily_returns will calculate returns. Gets passed to an xts, so simply use xts subsetting notation. If missing or NULL, then the full date range of OHLCV is used.

buy_at
Character: Specifies the 'initial price' to be used for calculating returns. Must be One of "open", "high", "low", "close", or "price". Case sensitive.

sell_at
Character: Specifies the 'final price' to be used for calculating returns. Must be One of "open", "high", "low", "close", or "price". Case sensitive.

include_volume
Boolean: include volume data in output? Default TRUE.

include_buy_sell
Boolean: include in the output the buy & sell prices that were used to calculate each return? Default FALSE.

returns_method
Character vector specifying how the returns should be calculated. Choices are: "ln", "log2", "log10", "pct_diff", and "multiple". Selecting multiple choices is allowed. Defaults to all.

Value

depends on value passed to include_volume:

DEFAULT, include_volume == FALSE: returns an xts object. Each element is the daily log return -- taking into account splits and dividends -- of the asset specified in the column name, realized on the date specified in the xts index relative to the previous trading day.

include_volume == TRUE: returns a list of two xts objects named "daily_returns" and "daily_volume". "daily_returns" is exactly the same xts object that is returned if include_volume == TRUE, as described above. "daily_volume" has the same structure, date indices, and colnames "daily_returns", but contains the trade volume observed for a given asset on a given day.