Calculate Historical Time-series Returns

Calculate period-over-period historical returns for a list of assets. calculate_historical_returns() takes dividends, splits, and mergers & acquisitions into account, as well as short fees if returns on short positions are to be included in the output. calculate_historical_returns is intended for use in calculating correlations, averages, and volatility of returns; therefore, it does not take taxes or transaction fees into account.

calculate_historical_returns(
  assets,
  date_range_xts = paste0("2012/", as.character(Sys.Date())),
  buy_at = "Close",
  sell_at = "Close",
  returns_method = "ln",
  short_fees = NULL,
  silent = FALSE
)

Arguments

assets
a list of sub-lists named and corresponding to a given asset, for example, "AAPL". Each sublist contains split, dividend, and aggregated prices data as xts objects. An example of an object that may be passed as asset_data is stock_data.

date_range_xts
A string that specifies the date range for which calculate_daily_returns() will calculate returns. Use xts subsetting notation to specify the date range; see examples in xts. If missing or NULL, then the full date range of OHLCV prices is used.

buy_at
Character: Specifies the 'initial price' to be used for calculating returns. Must be one of "open", "high", "low", "close", or "price", defaults to "Close". Not case sensitive.

sell_at
Character: Specifies the 'final price' to be used for calculating returns. Must be One of "open", "high", "low", "close", or "price", defaults to "Close". Not case sensitive.

returns_method
Character vector specifying how the returns should be calculated, defaults to "ln. Choices are:

"ln": Log return; ln(sell_price / buy_price) where ln() denotes the natural logarithm (base e).

"log2": Log return; log(sell_price / buy_price) where log() denotes the logarithm to base 2.

"log10": Log return; log(sell_price / buy_price) where log() denotes the logarithm to base 10.

"pct_diff": Percent difference; (sell_price - buy_price)/buy_price. NOT multiplied by 100.

"multiple": Price multiple; sell_price / buy_price.

short_fees
Optional. Either a named numeric vector or a numeric length 1 (i.e., a single number). Specify this parameter to cause calculate_historical_returns() to include the returns observed for short selling each asset. calculate_historical_returns() will assume that, for every return reported in the output, the expected return on a short sale equals the expected return on the long sale times (1 - short fees): $$ \langle E_r,short \rangle = \langle E_r,long \rangle \times (short\_fees - 1) $$

If a single number (numeric vector of length 1): The short fees for every asset in assets are assumed to be equal to the number passed in as short_fees. Use this option if you expect the short fees for each asset to be the same.

If numeric vector length > 1: The names of short_fees must correspond to the names of assets, and each element must be the short fee expected for each asset. Use this option if the short fees for each asset is expected to be significantly different.

See the "The Cost of Shorting" section for more information.

silent
Boolean. If FALSE (Default), messages will be displayed for special cases that might produce unexpected behavior; e.g., if no data is available for the date range specified. If TRUE, these messages will not be displayed.

Value

An xts object. Each element is the return observed on the date given by the xts's index with respect to the previous period, taking into account any splits, dividends or M&A events that may have occurred in between, for the asset specified by the element's column name.

Examples

# Calculate daily natural log returns of Apple (AAPL) and AT&T (T) for 2014, a
# year in which Apple had a major stock split.
aapl_att_returns <- calculate_historical_returns(
  assets         = stock_data[c("AAPL", "T")],
  date_range_xts = "2014"
)

# Print the first 10 rows:
head(aapl_att_returns, 10)

#>                    AAPL             T
#> 2014-01-03 -0.022210746 -0.0043010819
#> 2014-01-06  0.005438253  0.0045871640
#> 2014-01-07 -0.007181981 -0.0002860821
#> 2014-01-08  0.006317525 -0.0071787817
#> 2014-01-09 -0.012854131 -0.0206557949
#> 2014-01-10 -0.006693129  0.0023823716
#> 2014-01-13  0.005221455 -0.0095637310
#> 2014-01-14  0.019702704  0.0053908486
#> 2014-01-15  0.019878344  0.0092166551
#> 2014-01-16 -0.005595502  0.0050184607

# Let's check this result in a few key places. 

# 1) Normal trading day
#      Because we didn't specify otherwise, calculate_historical_returns() will
#      use its default behavior and will calculate period-over-period
#      natural-log returns using Close prices. Let's pick Monday, 21 July 2014,
#      which was a normal trading day (no dividend, no split) for AT&T. The
#      previous trading day was Friday, 18 Jul 2014:
stock_data$T$prices["2014-07-18/2014-07-21"]

#>             Open  High   Low Close    TWAP    VWAP   Volume PreviousCloseDate
#> 2014-07-18 36.18 36.19 36.00 36.17 36.1052 36.1115 16132950             16268
#> 2014-07-21 36.07 36.16 35.88 35.96 36.0222 36.0169 20917407             16269
#>            PreviousClose
#> 2014-07-18         36.03
#> 2014-07-21         36.17

#      The natural log return using Close prices (calculate_historical_returns()
#      default) as calculated by hand should be:
log(
  as.numeric(stock_data$T$prices$Close["2014-07-21"]) /
    as.numeric(stock_data$T$prices$Close["2014-07-18"])
)

#> [1] -0.005822836

#      This return was realized on 2014-07-21, so it should appear at that date
#      index in the results returned by calculate_historical_returns():
aapl_att_returns["2014-07-21", "T"]

#>                       T
#> 2014-07-21 -0.005822836

# 2) AT&T's dividend
#      AT&T paid several dividends in 2014:
stock_data$T$dividends["2014"]

#>            DividendAmount RecordDate   DeclaredDate Type              
#> 2014-01-08 "0.46000"      "2014-01-10" "2013-12-13" "OrdinaryDividend"
#> 2014-04-08 "0.46000"      "2014-04-10" "2014-03-28" "OrdinaryDividend"
#> 2014-07-08 "0.46000"      "2014-07-10" "2014-06-27" "OrdinaryDividend"
#> 2014-10-08 "0.46000"      "2014-10-10" "2014-09-26" "OrdinaryDividend"
#>            PaymentFrequency
#> 2014-01-08 "Quarterly"     
#> 2014-04-08 "Quarterly"     
#> 2014-07-08 "Quarterly"     
#> 2014-10-08 "Quarterly"     

#     If you had bought a share of AT&T stock on 07 July 2014 and sold it at the
#     Close price the next day after the dividend went ex-div, you would have
#     collected the closing price from the sale plus the dividend amount, 
#     meaning that your total earned return would have been:
log(
  (
    as.numeric(stock_data$T$prices$Close["2014-07-08"]) +
      as.numeric(stock_data$T$dividends$DividendAmount["2014-07-08"])
  ) / as.numeric(stock_data$T$prices$Close["2014-07-07"])
)

#> [1] 0

#      This return was realized on 2014-07-08, so it should appear at that date
#      index in the results returned by calculate_historical_returns():
aapl_att_returns["2014-07-08", "T"]

#>            T
#> 2014-07-08 0

# --> In this case, the stock closed at the same price as the previous day once
#     the dividend was taken into account, so return = 0.

# 3) AAPL's split
#      Apple underwent a major split on a Monday in June of 2014:
stock_data$AAPL$splits

#>            Denominator Numerator
#> 2014-06-09           7         1

#     If you had bought a share of AAPL at Close price on Friday, you would have
#     woken up on Monday morning with 7 shares... each one of which worth about
#     1/7th what they were worth previously. If you sold those shares at Close
#     price on Monday, your total return would be:
log(
  (
    as.numeric(stock_data$AAPL$prices$Close["2014-06-09"]) *
      as.numeric(stock_data$AAPL$splits$Denominator["2014-06-09"])
  ) / as.numeric(stock_data$AAPL$prices$Close["2014-06-06"])
)

#> [1] 0.01587469

#      This return was realized on 2014-06-09, so it should appear at that date
#      index in the results returned by calculate_historical_returns():
aapl_att_returns["2014-06-09", "AAPL"]

#>                  AAPL
#> 2014-06-09 0.01587469

# 4) Shorts
#   a) Get returns for General Dynamics (GD), Chubb Ltd. (CB), and iShares' Core
#        S&P 500 index (IVV), including shorts, assuming the following short 
#        fees for each of these assets:
gd_cb_ivv_returns_w_shorts <- calculate_historical_returns(
  assets         = stock_data[c("GD", "CB", "IVV")],
  date_range_xts = "2020-08-05/2020-08-14",
  short_fees     = c("CB" = 0.004, "GD" = 0.0025, "IVV" = 0.0025)
)
gd_cb_ivv_returns_w_shorts

#>                      CB            GD           IVV         s_CB          s_GD
#> 2020-08-06  0.007057080 -0.0003351993  0.0067547630 -0.007028852  0.0003343613
#> 2020-08-07  0.026433257  0.0253563616  0.0007146473 -0.026327524 -0.0252929707
#> 2020-08-10  0.007742172  0.0231974994  0.0030018896 -0.007711203 -0.0231395056
#> 2020-08-11  0.007985744  0.0145210376 -0.0084041621 -0.007953801 -0.0144847350
#> 2020-08-12 -0.013191697 -0.0154795960  0.0138498959  0.013138930  0.0154408970
#> 2020-08-13  0.007036896 -0.0093782805 -0.0017133915 -0.007008749  0.0093548348
#> 2020-08-14 -0.008112082  0.0096979020 -0.0000887036  0.008079634 -0.0096736573
#>                    s_IVV
#> 2020-08-06 -6.737876e-03
#> 2020-08-07 -7.128607e-04
#> 2020-08-10 -2.994385e-03
#> 2020-08-11  8.383152e-03
#> 2020-08-12 -1.381527e-02
#> 2020-08-13  1.709108e-03
#> 2020-08-14  8.848184e-05

#   b) Now assume that the short fee for each stock is the same (0.25%):
gd_cb_ivv_returns_w_shorts <- calculate_historical_returns(
  assets         = stock_data[c("GD", "CB", "IVV")],
  date_range_xts = "2020-08-05/2020-08-14",
  short_fees     = 0.0025
)
gd_cb_ivv_returns_w_shorts

#>                      CB            GD           IVV         s_CB         s_GD
#> 2020-08-06  0.007057080 -0.0003351993  0.0067547630 -0.004557080  0.002835199
#> 2020-08-07  0.026433257  0.0253563616  0.0007146473 -0.023933257 -0.022856362
#> 2020-08-10  0.007742172  0.0231974994  0.0030018896 -0.005242172 -0.020697499
#> 2020-08-11  0.007985744  0.0145210376 -0.0084041621 -0.005485744 -0.012021038
#> 2020-08-12 -0.013191697 -0.0154795960  0.0138498959  0.015691697  0.017979596
#> 2020-08-13  0.007036896 -0.0093782805 -0.0017133915 -0.004536896  0.011878280
#> 2020-08-14 -0.008112082  0.0096979020 -0.0000887036  0.010612082 -0.007197902
#>                    s_IVV
#> 2020-08-06 -0.0042547630
#> 2020-08-07  0.0017853527
#> 2020-08-10 -0.0005018896
#> 2020-08-11  0.0109041621
#> 2020-08-12 -0.0113498959
#> 2020-08-13  0.0042133915
#> 2020-08-14  0.0025887036