Calculate Market Portfolio

Calculate the Sharpe-optimal market portfolio available for a set of assets given the expected returns, volatilities, and correlations of returns for each asset.

calculate_market_portfolio(
  exp_rtn,
  exp_vol,
  exp_cor,
  rfr = 2.7397e-05,
  allow_shorts = FALSE,
  prices = NULL,
  portfolio_aum = NULL
)

Arguments

exp_rtn
Named numeric vector for which each element is the return expected for the asset specified by the element's name.

exp_vol
Named numeric vector for which each element is the volatility expected for the asset specified by the element's name.

exp_cor
Named numeric matrix specifying the expected covariance of returns for each asset pair.

rfr
The risk-free rate (in decimal form; i.e., to specify a rate of "3%" use "0.03", not "3") that YOU can earn on cash with reasonable liquidity constraints that YOU can tolerate. For big banks and in economic textbooks, this rate is usually the current rate on 3-month T-bills (or even higher). That might be just fine for you if you plan on rebalancing once a year. For a trader running a strategy that can't tolerate cash being tied up for 3 months at a time, the rfr should be set to whatever interest rate your brokerage is giving you on cash in your trading account. Used in calculation of the Sharpe Ratio. Defaults to 0.0027397% daily return (about 1% annually). If you specify a different value for rfr, make sure its time basis matches the one used for you other inputs, (i.e., if exp_rtn contains monthly returns, use monthly risk-free rate)!

allow_shorts
Defaults to FALSE; set to TRUE to allow shorting of all the assets. There are MANY ways to do this, but by default calculate_market_portfolio() simply treats shorts as another asset whose expected return equals negative the expected return of going long.

prices
Optional: a named numeric vector for which each name is the identifier of an asset and each element is the current price of that asset for which a market portfolio is to be calculated on a shares basis. See the "Returns" section for more info.

portfolio_aum
Optional: numeric, length 1, giving the total amount of assets under management for which a market portfolio is to be calculated on a shares basis. See "Returns" section for more info.

Value

If weights basis (prices and portfolio_aum not specified): #' A list describing the market portfolio (MP). Contains the following four elements:

sharpe, numeric:

Sharpe Ratio of MP

weights, named numeric vector:

The values of weights range from 0 to 1 and denote the fraction of the total MP value allocated to the asset whose identifier is that value's name.

ex_return, numeric:

The return that is expected (i.e., predicted/forecast) for the MP over the next time interval.

ex_volatility:

The volatility that is expected (i.e., predicted/forecast) for the MP over the next time interval.

If shares basis (prices and portfolio_aum specified): calculate_market_portfolio() will optimize the portfolio assuming that portfolio_aum is available for investing and that the price of each asset is given by prices. Returns a list describing the market portfolio (MP) which contains the for elements described above, plus three more:

shares:

Named numeric vector giving the number of shares of each asset calculated for the market portfolio

prices:

Named umeric vector giving the prices used by calculate_market_portfolio() to determine the MP.

cash:

Numeric, length 1. Leftover cash not invested in assets.

Details

All arguments which are percentages (exp_rtn, exp_vol, and rfr) must be supplied in decimal form; i.e., to specify "12%", use 0.12, not 12.

It should go without saying, but make sure that every asset is represented in the three inputs: names of exp_rtn, exp_vol, and the row & colnames of exp_cor.

This function works by finding the Sharpe-optimum equal-weighted portfolio that can be created using the assets passed in. Using that portfolio as a starting point, the function finds the assets A and B such that reallocating ("selling") a small amount (step) of A into asset B ("buying") results in a portfolio whose Sharpe is greater than all other possible assets A and B. The portfolio variable is updated.

This process is repeated with smaller and smaller step sizes. When it is not possible to reallocate step amount of any asset A into any other asset B so as to create a portfolio having a better Sharpe, the Market Portfolio has been reached and the function returns the value.

Weights mode: If either prices or portfolio_aum is unspecified in the call to calculate_market_portfolio(), then the function operates in "Weights Mode" and will attempt to answer the question: "What is the the optimum fractional, real-valued weight of each asset in the Sharpe-optimal MP?" Because the Sharpe ratio and the optimum weights are real numbers, this operation can be computed with arbitrary precision, to infinte decimal places. calculate_market_portfolio() was written to be clearly understood and run reasonably fast, and testing indicates that the weights are reliable to within +/- 1% (i.e., "0.18" to "0.20" for a reported weight of "0.19") of the optimal value. As such, weights reported by weights mode should be considered somewhat approximate.

Shares mode: A much different situation arises when we ask the question: "Given $5MM in capital, what's the optimum portfolio that can be built from a set of assets trading at given prices?" Note that the formulation of this question requires both prices and portfolio_aum to be provided. Because shares can only be bought in integer values, and because the sum of the capital allocated to each asset (including "cash") must equal portfolio_aum, the number of values available for the shares vector is countably finite and it is possible to find an exact solution.

Examples

# Let's pick an example date for which to create an MP:
mp_date <- "2018-01-09"
# Feel free to change this date to any other date for which you have data.

# Use calculate_returns() to get the daily returns observed for the 365 days
# ending on mp_date:
historical_rtn <- calculate_returns(
  assets         = stock_data,
  date_range_xts = paste0(
    as.Date(mp_date) - 365,
    "/",
    mp_date
  )
)

# We'll assume that the return we expect over the next year is the annualized 
# GMMR of the daily rates of return in historical_rtn:
exp_rtn <- gmrr(historical_rtn)

# Assume that the volatilities we expect for the next year are the annualized
# daily vols we observed during the previous year:
exp_vol <- dplyr::summarize(
  tibble::as_tibble(historical_rtn),
  dplyr::across(dplyr::everything(), sd, na.rm = TRUE)
) %>%
  purrr::as_vector()

# Assume that the correlations of returns of each asset pair that we expect
# for the next year will be the same as the previous year:
exp_cor <- stats::cor(historical_rtn, use = "pairwise.complete.obs")

# Calculate the market portfolio:
mp_by_wt <- calculate_market_portfolio(exp_rtn, exp_vol, exp_cor)
mp_by_wt

#> $sharpe
#> [1] 0.3321399
#> 
#> $weights
#>        AAPL        ABBV         ADM         AFL         AOS         ATO 
#> 0.062692308 0.105153846 0.000000000 0.100692308 0.000000000 0.000000000 
#>          CB          CL         COP        CTAS         CVX         EMR 
#> 0.000000000 0.000000000 0.000000000 0.002769231 0.028461538 0.000000000 
#>         ESS         FRT          GD         HRL         IVV         JNJ 
#> 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000 0.024461538 
#>         KMB         LEG         MCD         MMM        MSFT         NSC 
#> 0.000000000 0.000000000 0.183923077 0.136615385 0.000000000 0.000000000 
#>         PEP         PNR         PSX         SHW           T         TFC 
#> 0.060153846 0.000000000 0.000000000 0.055230769 0.000000000 0.000000000 
#>        TROW         TXN         VFC         WBA         WMT 
#> 0.066153846 0.028461538 0.055923077 0.000000000 0.089307692 
#> 
#> $ex_return
#> [1] 0.001405307
#> 
#> $ex_volatility
#> [1] 0.004148584
#> 

### Calculate the market portfolio allowing both long & short positions:
mp_by_wt_shorts <- calculate_market_portfolio(
  exp_rtn, exp_vol, exp_cor, allow_shorts = TRUE
)
mp_by_wt_shorts

#> $sharpe
#> [1] 0.401824
#> 
#> $weights
#>        AAPL        ABBV         ADM         AFL         AOS         ATO 
#> 0.037157895 0.047894737 0.000000000 0.072105263 0.006315789 0.000000000 
#>          CB          CL         COP        CTAS         CVX         EMR 
#> 0.007368421 0.000000000 0.002631579 0.025789474 0.032947368 0.023473684 
#>         ESS         FRT          GD         HRL         IVV         JNJ 
#> 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000 0.017684211 
#>         KMB         LEG         MCD         MMM        MSFT         NSC 
#> 0.000000000 0.000000000 0.074947368 0.061263158 0.000000000 0.000000000 
#>         PEP         PNR         PSX         SHW           T         TFC 
#> 0.063684211 0.000000000 0.000000000 0.043368421 0.000000000 0.000000000 
#>        TROW         TXN         VFC         WBA         WMT      s_AAPL 
#> 0.043157895 0.012105263 0.036631579 0.000000000 0.048947368 0.000000000 
#>      s_ABBV       s_ADM       s_AFL       s_AOS       s_ATO        s_CB 
#> 0.000000000 0.032421053 0.000000000 0.000000000 0.010947368 0.000000000 
#>        s_CL       s_COP      s_CTAS       s_CVX       s_EMR       s_ESS 
#> 0.002105263 0.000000000 0.000000000 0.000000000 0.000000000 0.036315789 
#>       s_FRT        s_GD       s_HRL       s_IVV       s_JNJ       s_KMB 
#> 0.028736842 0.032631579 0.016315789 0.024631579 0.000000000 0.011578947 
#>       s_LEG       s_MCD       s_MMM      s_MSFT       s_NSC       s_PEP 
#> 0.034736842 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000 
#>       s_PNR       s_PSX       s_SHW         s_T       s_TFC      s_TROW 
#> 0.030000000 0.005263158 0.000000000 0.005789474 0.036631579 0.000000000 
#>       s_TXN       s_VFC       s_WBA       s_WMT 
#> 0.000000000 0.000000000 0.034421053 0.000000000 
#> 
#> $ex_return
#> [1] 0.0007847404
#> 
#> $ex_volatility
#> [1] 0.001884764
#> 

### Repeat the above, assuming you have $250,000 to invest and that the stocks
### may be bought at their closing prices on mp_date:

prices <- stock_data %>%
  vapply(
    function(stock){
      stock$prices$Close[mp_date]
    },
    numeric(1)
  )
portfolio_aum <- 250000

mp_by_shares <- calculate_market_portfolio(
  exp_rtn,
  exp_vol,
  exp_cor,
  prices        = prices,
  portfolio_aum = portfolio_aum
)
mp_by_shares

#> $sharpe
#> [1] 0.3321391
#> 
#> $shares
#> AAPL ABBV  ADM  AFL  AOS  ATO   CB   CL  COP CTAS  CVX  EMR  ESS  FRT   GD  HRL 
#>   90  261    0  280    0    0    0    0    0    4   55    0    0    0    0    0 
#>  IVV  JNJ  KMB  LEG  MCD  MMM MSFT  NSC  PEP  PNR  PSX  SHW    T  TFC TROW  TXN 
#>    0   42    0    0  264  141    0    0  127    0    0   32    0    0  147   64 
#>  VFC  WBA  WMT 
#>  183    0  222 
#> 
#> $cash
#> [1] 1101.2
#> 
#> $weights
#>       AAPL       ABBV        ADM        AFL        AOS        ATO         CB 
#> 0.06275880 0.10465056 0.00000000 0.10080000 0.00000000 0.00000000 0.00000000 
#>         CL        COP       CTAS        CVX        EMR        ESS        FRT 
#> 0.00000000 0.00000000 0.00257952 0.02812480 0.00000000 0.00000000 0.00000000 
#>         GD        HRL        IVV        JNJ        KMB        LEG        MCD 
#> 0.00000000 0.00000000 0.00000000 0.02421552 0.00000000 0.00000000 0.18325824 
#>        MMM       MSFT        NSC        PEP        PNR        PSX        SHW 
#> 0.13608192 0.00000000 0.00000000 0.05998464 0.00000000 0.00000000 0.05429888 
#>          T        TFC       TROW        TXN        VFC        WBA        WMT 
#> 0.00000000 0.00000000 0.06570900 0.02825984 0.05572716 0.00000000 0.08914632 
#> 
#> $ex_return
#> [1] 0.001399086
#> 
#> $ex_volatility
#> [1] 0.004129863
#> 
#> $prices
#>   AAPL   ABBV    ADM    AFL    AOS    ATO     CB     CL    COP   CTAS    CVX 
#> 174.33 100.24  40.22  90.00  63.24  82.59 145.85  75.64  57.93 161.22 127.84 
#>    EMR    ESS    FRT     GD    HRL    IVV    JNJ    KMB    LEG    MCD    MMM 
#>  72.32 236.27 125.44 205.88  36.40 276.57 144.14 116.80  47.95 173.54 241.28 
#>   MSFT    NSC    PEP    PNR    PSX    SHW      T    TFC   TROW    TXN    VFC 
#>  88.22 153.60 118.08  72.88 101.72 424.21  37.49  52.09 111.75 110.39  76.13 
#>    WBA    WMT 
#>  73.42 100.39 
#> 

### Allow shorting:
mp_by_shares_shorts <- calculate_market_portfolio(
  exp_rtn,
  exp_vol,
  exp_cor,
  allow_shorts  = TRUE,
  prices        = prices,
  portfolio_aum = portfolio_aum
)
mp_by_shares_shorts

#> $sharpe
#> [1] 0.401821
#> 
#> $shares
#>   AAPL   ABBV    ADM    AFL    AOS    ATO     CB     CL    COP   CTAS    CVX 
#>     53    119      0    200     24      0     12      0     11     40     64 
#>    EMR    ESS    FRT     GD    HRL    IVV    JNJ    KMB    LEG    MCD    MMM 
#>     81      0      0      0      0      0     30      0      0    107     63 
#>   MSFT    NSC    PEP    PNR    PSX    SHW      T    TFC   TROW    TXN    VFC 
#>      0      0    134      0      0     25      0      0     96     27    120 
#>    WBA    WMT s_AAPL s_ABBV  s_ADM  s_AFL  s_AOS  s_ATO   s_CB   s_CL  s_COP 
#>      0    121      0      0    201      0      0     33      0      6      0 
#> s_CTAS  s_CVX  s_EMR  s_ESS  s_FRT   s_GD  s_HRL  s_IVV  s_JNJ  s_KMB  s_LEG 
#>      0      0      0     38     57     39    112     21      0     25    180 
#>  s_MCD  s_MMM s_MSFT  s_NSC  s_PEP  s_PNR  s_PSX  s_SHW    s_T  s_TFC s_TROW 
#>      0      0      0      0      0    103     13      0     39    175      0 
#>  s_TXN  s_VFC  s_WBA  s_WMT 
#>      0      0    116      0 
#> 
#> $cash
#> [1] 2144.83
#> 
#> $weights
#>       AAPL       ABBV        ADM        AFL        AOS        ATO         CB 
#> 0.03695796 0.04771424 0.00000000 0.07200000 0.00607104 0.00000000 0.00700080 
#>         CL        COP       CTAS        CVX        EMR        ESS        FRT 
#> 0.00000000 0.00254892 0.02579520 0.03272704 0.02343168 0.00000000 0.00000000 
#>         GD        HRL        IVV        JNJ        KMB        LEG        MCD 
#> 0.00000000 0.00000000 0.00000000 0.01729680 0.00000000 0.00000000 0.07427512 
#>        MMM       MSFT        NSC        PEP        PNR        PSX        SHW 
#> 0.06080256 0.00000000 0.00000000 0.06329088 0.00000000 0.00000000 0.04242100 
#>          T        TFC       TROW        TXN        VFC        WBA        WMT 
#> 0.00000000 0.00000000 0.04291200 0.01192212 0.03654240 0.00000000 0.04858876 
#>     s_AAPL     s_ABBV      s_ADM      s_AFL      s_AOS      s_ATO       s_CB 
#> 0.00000000 0.00000000 0.03233688 0.00000000 0.00000000 0.01090188 0.00000000 
#>       s_CL      s_COP     s_CTAS      s_CVX      s_EMR      s_ESS      s_FRT 
#> 0.00181536 0.00000000 0.00000000 0.00000000 0.00000000 0.03591304 0.02860032 
#>       s_GD      s_HRL      s_IVV      s_JNJ      s_KMB      s_LEG      s_MCD 
#> 0.03211728 0.01630720 0.02323188 0.00000000 0.01168000 0.03452400 0.00000000 
#>      s_MMM     s_MSFT      s_NSC      s_PEP      s_PNR      s_PSX      s_SHW 
#> 0.00000000 0.00000000 0.00000000 0.00000000 0.03002656 0.00528944 0.00000000 
#>        s_T      s_TFC     s_TROW      s_TXN      s_VFC      s_WBA      s_WMT 
#> 0.00584844 0.03646300 0.00000000 0.00000000 0.00000000 0.03406688 0.00000000 
#> 
#> $ex_return
#> [1] 0.0007799149
#> 
#> $ex_volatility
#> [1] 0.001872769
#> 
#> $prices
#>   AAPL   ABBV    ADM    AFL    AOS    ATO     CB     CL    COP   CTAS    CVX 
#> 174.33 100.24  40.22  90.00  63.24  82.59 145.85  75.64  57.93 161.22 127.84 
#>    EMR    ESS    FRT     GD    HRL    IVV    JNJ    KMB    LEG    MCD    MMM 
#>  72.32 236.27 125.44 205.88  36.40 276.57 144.14 116.80  47.95 173.54 241.28 
#>   MSFT    NSC    PEP    PNR    PSX    SHW      T    TFC   TROW    TXN    VFC 
#>  88.22 153.60 118.08  72.88 101.72 424.21  37.49  52.09 111.75 110.39  76.13 
#>    WBA    WMT s_AAPL s_ABBV  s_ADM  s_AFL  s_AOS  s_ATO   s_CB   s_CL  s_COP 
#>  73.42 100.39 174.33 100.24  40.22  90.00  63.24  82.59 145.85  75.64  57.93 
#> s_CTAS  s_CVX  s_EMR  s_ESS  s_FRT   s_GD  s_HRL  s_IVV  s_JNJ  s_KMB  s_LEG 
#> 161.22 127.84  72.32 236.27 125.44 205.88  36.40 276.57 144.14 116.80  47.95 
#>  s_MCD  s_MMM s_MSFT  s_NSC  s_PEP  s_PNR  s_PSX  s_SHW    s_T  s_TFC s_TROW 
#> 173.54 241.28  88.22 153.60 118.08  72.88 101.72 424.21  37.49  52.09 111.75 
#>  s_TXN  s_VFC  s_WBA  s_WMT 
#> 110.39  76.13  73.42 100.39 
#>