Market Portfolio (MP) from Historical Returns

Given an xts, calculate the

market_portfolio_from_hist_rtn(rtn_xts, rf = 0.0025)

Arguments

rtn_xts
An xts xts object whose values are period-over-period returns observed for the assets identified by column names, and for which the market portfolio is sought.

rf
The risk-free rate to use in calculation of Sharpe Ratio, in decimal form. Defaults to 0.0025% daily return (about 0.91% annually). If you specify a different value for rf, make sure its units match the returns in rtn_xts, (i.e., if rtn_xts contains monthly returns, use monthly risk-free rate) and don't forget to use decimal representation (i.e., rf = 5 is taken to mean 500%, not 5%).

Value

A list describing the market portfolio (MP) found for rtn_xts, having four elements:

sharpe, numeric:

Sharpe Ratio of MP

weights, named numeric vector:

The values of weights range from 0 to 1 and denote the fraction of the total MP value allocated to the asset whose identifier is that value's name.

ex_return, numeric:

The return that is expected (i.e., predicted/forecast) for the MP over the next time interval.

ex_volatility:

The volatility that is expected (i.e., predicted/forecast) for the MP over the next time interval.

Details

This function works by finding the Sharpe-optimum equal-weighted portfolio that can be created using the assets passed in. Using that portfolio as a starting point, the function finds the assets A and B such that reallocating ("selling") a small amount (step) of A into asset B ("buying") results in a portfolio whose Sharpe is greater than all other possible assets A and B. The portfolio variable is updated.

This process is repeated with smaller and smaller step sizes. When it is not possible to reallocate step amount of any asset A into any other asset B so as to create a portfolio having a better Sharpe, the Market Portfolio has been reached and the function returns the value.