Exchange Hours

Operating hours and related information for worldwide exchanges. Downloaded from Xe's World Currency Symbols and stored as a tibble.

stock_exchange_hours

Format

An object of class tbl_df (inherits from tbl, data.frame) with 70 rows and 13 columns.

Examples

stock_exchange_hours

#> Warning: `...` is not empty.
#> 
#> We detected these problematic arguments:
#> * `needs_dots`
#> 
#> These dots only exist to allow future extensions and should be empty.
#> Did you misspecify an argument?

#> # A tibble: 70 x 13
#>    Name  ID    Country City  Zone  time_delta DST   local_Open local_Close
#>    <chr> <chr> <chr>   <chr> <chr> <chr>      <chr> <chr>      <chr>      
#>  1 Euro… AMS   Nether… Amst… CET   +1         "Mar… 09:00      17:40      
#>  2 Arme… AMX   Armenia Yere… AMT   +4         ""    10:00      15:00      
#>  3 Amma… ASE   Jordan  Amman EET   +2         "Mar… 10:00      12:00      
#>  4 Aust… ASX   Austra… Sydn… AEST  +10        "Oct… 10:00      16:00      
#>  5 Buen… BCBA  Argent… Buen… ART   −3         ""    11:00      17:00      
#>  6 Span… BME   Spain   Madr… CET   +1         "Mar… 09:00      17:30      
#>  7 Mexi… BMV   Mexico  Mexi… CST   −6         "Apr… 08:30      15:00      
#>  8 Bols… Bove… Brazil  São … BRT   −3         "Sus… 10:00      17:30      
#>  9 Euro… BRU   Belgium Brus… CET   +1         "Mar… 09:00      17:30      
#> 10 Bomb… BSE   India   Mumb… IST   +5.5       ""    09:15      15:30      
#> # … with 60 more rows, and 4 more variables: local_Lunch <chr>, UTC_Open <chr>,
#> #   UTC_Close <chr>, UTC_Lunch <chr>