Stock Info

Get all available information about a stock printed cleanly as a tibble. stock_info() is not a data object; rather, it's a function that scans a data object (stock_data) and compiles and returns the gathered information in a readable format.

stock_info(stocks = stock_data, symbols = names(stocks))

Arguments

stocks

List of stocks following the same structure as stock_data. Defaults to stock_data.

symbols

Character vector containing symbols to be included in the output. Each element of symbols must appear as a name of a stock element in stocks. Defaults to names(stocks).

Value

A tibble containing available descriptive data for each member of stock_data. Not case-sensitive.

Disclaimer

Absolutely no promise, guarantee, or representation is made by the authors of FinancieR concerning the accuracy or reliability of stock_data, or any other data in the FinancieR package.

Examples

  stock_info()
#> # A tibble: 10 × 14
#>    Symbol Name             CompositeFIGI Industry Sector MarketIdentific… Market
#>    <chr>  <chr>            <chr>         <chr>    <chr>  <chr>            <chr> 
#>  1 AAPL   Apple Inc        BBG000B9XRY4  Telecom… Elect… XNAS             NASDAQ
#>  2 T      AT&T, Inc.       BBG000BSJK37  MajorTe… Commu… XNYS             New Y…
#>  3 PX     Praxair Inc      NA            Unknown  Unkno… XNYS             New Y…
#>  4 LIN    Linde Plc        BBG00GVR8YQ9  Chemica… Proce… XNYS             New Y…
#>  5 IVV    BlackRock Insti… BBG000BVZ4F5  Investm… Misce… ARCX             NYSE …
#>  6 TXN    Texas Instrumen… BBG000BVV7G1  Semicon… Elect… XNAS             NASDAQ
#>  7 PEP    Pepsico Inc.     BBG000DH7JK6  Beverag… Consu… XNAS             NASDAQ
#>  8 CB     Chubb Limited    BBG000BR14K5  Propert… Finan… XNYS             New Y…
#>  9 GD     General Dynamic… BBG000BK67C7  Aerospa… Elect… XNYS             New Y…
#> 10 ABT    Abbott Laborato… BBG000B9ZXB4  Medical… Healt… XNYS             New Y…
#> # … with 7 more variables: InstrumentClass <chr>, FIGI <chr>, Valoren <chr>,
#> #   CIK <chr>, SEDOL <chr>, ISIN <chr>, CUSIP <chr>