Using [ on Objects of Class "stock"

You can quickly retrieve price, dividend, split, and MnA data useing the [ operator on an object of class stock, much like subsetting a matrix or list in base R.

# S3 method for stock
[(x, i, j = NULL, silent = FALSE)

Arguments

x
an object of class "stock".

i
an xts subsetting string following the same rules as for an xts object: "2015" will return data for all of 2015, "2018-04" will return data for March of 2018 only, "2016-09-19/2016-12-31" will return data from 19 Sep 2016 to the end of the year, etc. If NULL, then the full time range of available data for the selected stock will be returned.

j
Character vector specifying column names to be included in the output. Available column names are:

"DividendAmount", "RecordDate", "DeclaredDate", "Type", "PaymentFrequency", "Open", "High", "Low", "Close", "TWAP", "VWAP", "Volume", "PreviousCloseDate", "PreviousClose", "Denominator", "Numerator", "ExDate"

silent
Boolean. If FALSE (Default), messages will be displayed for special cases that might produce unexpected behavior; e.g., if no data is available for the date range specified. If TRUE, these messages will not be displayed.

Value

an xts object, or NULL if no data is available for the specified asset within the specified date range.

Details

The rules are:

If a dividend took place during the specified date range, then the column "DividendAmount" will always be included in the output.

If a split took place during the specified date range, then the columns "Denominator" and "Numerator" will always be included in the output. "Denominator" refers to the number of shares AFTER the split and "Numerator" number of shares BEFORE the split; for example, a 2-to-1 split would have "Denominator" = 2 and "Numerator" = 1.

If an M&A event took place during the specified time range resulting in a change in the stock's ticker symbol, then the columns "symbol" and "multiple" will always be included in the output. "symbol" is the ticker symbol that the company traded under through time, and "multiple" gives the ratio number of shares held prior to MnA / number of shares held after the MnA.

Examples

# If no columns are specified, all available columns are returned.
# Single date:
stock_data$IVV["2019-06-11"]

#>            Open   High    Low  Close     TWAP     VWAP  Volume
#> 2019-06-11  293 293.42 290.18 290.95 291.3605 291.3883 2923455
#>            PreviousCloseDate PreviousClose
#> 2019-06-11             18057           291

# Dividends are correctly applied, even when issued on a non-trading day. In
# 2012, for example, Texas Instruments (TXN) had scheduled a dividend for the
# 29th of October but the exchange was closed due to Hurricane Sandy. The
# dividend, then, should be applied to the next trading day (31 October). Here,
# we'll select just the closing price for a date range around that time.
# Remember, any dividends that occur within the date range are always included.
stock_data$TXN["2012-10-25/2012-11-05", "Close"]

#>            Close DividendAmount
#> 2012-10-25 28.13           0.00
#> 2012-10-26 28.92           0.00
#> 2012-10-31 28.09           0.21
#> 2012-11-01 29.21           0.00
#> 2012-11-02 28.53           0.00
#> 2012-11-05 29.47           0.00

# Mergers are handled as illustrated below for the PX-LIN merger in 2018.
# Shareholders of PX received 1 share of the new company, LIN, for every 1
# share of PX.
stock_data$PX["2018-10-29/2018-11-05", "Close"]

#>            Close    symbol multiplier
#> 2018-10-29 "165.49" "PX"   "1"       
#> 2018-10-30 "164.5"  "PX"   "1"       
#> 2018-10-31 "165.47" "LIN"  "1"       
#> 2018-11-01 "162"    "LIN"  "1"       
#> 2018-11-02 "159.84" "LIN"  "1"       
#> 2018-11-05 "161.99" "LIN"  "1"       

# Splits are handled as follows for Apple's 7-to-1 split of 2014:
stock_data$AAPL["2014-06-05/2014-06-11", c("Open", "Close")]

#>              Open  Close Numerator Denominator
#> 2014-06-05 646.20 647.35         1           1
#> 2014-06-06 649.90 645.57         1           1
#> 2014-06-09  92.70  93.70         1           7
#> 2014-06-10  94.73  94.25         1           1
#> 2014-06-11  94.13  93.86         1           1

# Not specifying which columns you want can quickly lead to a lot of data, as in
# the below subset of IVV for the month of June 2019:
stock_data$IVV["2019-06"]

#>            Open     High       Low        Close    TWAP       VWAP      
#> 2019-06-03 "277.15" "278.39"   "274.9"    "276.38" "276.4588" "276.4957"
#> 2019-06-04 "278.96" "282.5759" "278.47"   "282.44" "280.6717" "280.8257"
#> 2019-06-05 "284.25" "284.9"    "282.211"  "284.86" "283.8816" "283.852" 
#> 2019-06-06 "285.21" "287.495"  "284.49"   "286.76" "285.8646" "285.8638"
#> 2019-06-07 "287.9"  "290.83"   "287.7"    "289.67" "289.7764" "289.7672"
#> 2019-06-10 "291.36" "292.81"   "290.87"   "291"    "291.679"  "291.6477"
#> 2019-06-11 "293"    "293.42"   "290.18"   "290.95" "291.3605" "291.3883"
#> 2019-06-12 "290.66" "291.25"   "289.83"   "290.29" "290.368"  "290.3781"
#> 2019-06-13 "291.4"  "291.97"   "290.6279" "291.62" "291.2763" "291.2941"
#> 2019-06-14 "291.31" "291.93"   "290.4001" "291.25" "290.9889" "290.988" 
#> 2019-06-17 "289.75" "290.45"   "289.41"   "289.74" "289.9867" "289.9679"
#> 2019-06-18 "291.63" "293.8"    "291.25"   "292.45" "292.7379" "292.7035"
#> 2019-06-19 "292.79" "293.89"   "291.78"   "293.24" "292.8821" "292.9638"
#> 2019-06-20 "296.32" "296.57"   "293.8788" "296.02" "295.6598" "295.6786"
#> 2019-06-21 "295.79" "297.2"    "295.47"   "295.57" "296.1241" "296.0067"
#> 2019-06-24 "295.9"  "296.27"   "295.15"   "295.27" "295.6836" "295.6679"
#> 2019-06-25 "295.35" "295.41"   "292.3"    "292.41" "293.677"  "293.6548"
#> 2019-06-26 "293.42" "293.99"   "292.02"   "292.08" "292.8215" "292.789" 
#> 2019-06-27 "292.99" "293.73"   "292.61"   "293.28" "293.3466" "293.3313"
#> 2019-06-28 "294.28" "295.21"   "293.71"   "294.75" "294.2672" "294.3027"
#>            Volume    PreviousCloseDate PreviousClose DividendAmount
#> 2019-06-03 "5589736" "18047"           "277.3"       "0"           
#> 2019-06-04 "3598917" "18050"           "276.38"      "0"           
#> 2019-06-05 "4674039" "18051"           "282.44"      "0"           
#> 2019-06-06 "4294539" "18052"           "284.86"      "0"           
#> 2019-06-07 "3180251" "18053"           "286.76"      "0"           
#> 2019-06-10 "3862913" "18054"           "289.67"      "0"           
#> 2019-06-11 "2923455" "18057"           "291"         "0"           
#> 2019-06-12 "2310269" "18058"           "290.95"      "0"           
#> 2019-06-13 "2304977" "18059"           "290.29"      "0"           
#> 2019-06-14 "2000184" "18060"           "291.62"      "0"           
#> 2019-06-17 "2511178" "18061"           "291.25"      "1.777745"    
#> 2019-06-18 "4622740" "18064"           "289.74"      "0"           
#> 2019-06-19 "7606221" "18065"           "292.45"      "0"           
#> 2019-06-20 "4408953" "18066"           "293.24"      "0"           
#> 2019-06-21 "3628064" "18067"           "296.02"      "0"           
#> 2019-06-24 "3549268" "18068"           "295.57"      "0"           
#> 2019-06-25 "2975598" "18071"           "295.27"      "0"           
#> 2019-06-26 "4210219" "18072"           "292.41"      "0"           
#> 2019-06-27 "2722438" "18073"           "292.08"      "0"           
#> 2019-06-28 "5751367" "18074"           "293.28"      "0"           
#>            RecordDate   DeclaredDate Type               PaymentFrequency
#> 2019-06-03 ""           ""           ""                 ""              
#> 2019-06-04 ""           ""           ""                 ""              
#> 2019-06-05 ""           ""           ""                 ""              
#> 2019-06-06 ""           ""           ""                 ""              
#> 2019-06-07 ""           ""           ""                 ""              
#> 2019-06-10 ""           ""           ""                 ""              
#> 2019-06-11 ""           ""           ""                 ""              
#> 2019-06-12 ""           ""           ""                 ""              
#> 2019-06-13 ""           ""           ""                 ""              
#> 2019-06-14 ""           ""           ""                 ""              
#> 2019-06-17 "2019-06-18" "2019-06-14" "OrdinaryDividend" "Quarterly"     
#> 2019-06-18 ""           ""           ""                 ""              
#> 2019-06-19 ""           ""           ""                 ""              
#> 2019-06-20 ""           ""           ""                 ""              
#> 2019-06-21 ""           ""           ""                 ""              
#> 2019-06-24 ""           ""           ""                 ""              
#> 2019-06-25 ""           ""           ""                 ""              
#> 2019-06-26 ""           ""           ""                 ""              
#> 2019-06-27 ""           ""           ""                 ""              
#> 2019-06-28 ""           ""           ""                 ""