Merge 2 XTS Tables According to Next Non-Missing Value

For two XTS objects xts1 and xts2, let A be any datetime index that appears in index(xts2), falls between min(index(xts1)) and max(index(xts1)), but does not appear in index(xts1). Let B be the earliest-occurring datetime index in index(xts1) that occurs after A.

xts_merge_align_next(xts1, xts2) will return the left-join merge of xts1 and xts2 with one additional operation: for any/all A and B that occur for the XTS pair, the values at at index B for the columns in the merged XTS coming from xts2 equal the aggregated value (as specified; e.g., sum) of the values of the columns in xts2 in the range A:B.

xts_merge_align_next(xts1, xts2, agg_function, ...)

Arguments

xts1
An XTS object

xts2
An XTS object that may or may not have indices that fall within the range of xts1, but do not appear in xts1

agg_function
Any function that accepts a numeric vector as input and returns a single number as output; e.g., sum

...
Additional parameters to pass to agg_function