Calculate Market Portfolio

Calculate the Sharpe-optimal market portfolio available for a set of assets given the expected returns, volatilities, and correlations of returns for each asset.

calculate_market_portfolio(
  exp_rtn,
  exp_vol,
  exp_cor,
  rf = 8.22e-05,
  allow_shorts = FALSE
)

Arguments

exp_rtn
Named numeric vector for which each element is the return expected for the asset specified by the element's name.

exp_vol
Named numeric vector for which each element is the volatility expected for the asset specified by the element's name.

exp_cor
Named numeric matrix specifying the expected covariance of returns for each asset pair.

rf
The risk-free rate to use in calculation of Sharpe Ratio, in decimal form. Defaults to 0.00822% daily return (about 3% annually). If you specify a different value for rf, make sure its time basis matches the one used for you other inputs, (i.e., if exp_rtn contains monthly returns, use monthly risk-free rate)!

allow_shorts
Defaults to FALSE; set to TRUE to allow shorting of all the assets. There are MANY ways to do this, but by default calculate_market_portfolio() simply treats shorts as another asset whose expected return equals negative the expected return of going long.

Value

A list describing the market portfolio (MP). Has four elements:

sharpe, numeric:

Sharpe Ratio of MP

weights, named numeric vector:

The values of weights range from 0 to 1 and denote the fraction of the total MP value allocated to the asset whose identifier is that value's name.

ex_return, numeric:

The return that is expected (i.e., predicted/forecast) for the MP over the next time interval.

ex_volatility:

The volatility that is expected (i.e., predicted/forecast) for the MP over the next time interval.

If total_capital and prices are specified, then two more elements are included in the list output:

shares, named numeric:

Integer number of shares to be bought

cash_balance, named numeric vector:

The values of weights range from 0 to 1 and denote the fraction of the total MP value allocated to the asset whose identifier is that value's name.

ex_return, numeric:

The return that is expected (i.e., predicted/forecast) for the MP over the next time interval.

ex_volatility:

The volatility that is expected (i.e., predicted/forecast) for the MP over the next time interval.

Details

All arguments which are percentages (exp_rtn, exp_vol, and rf) must be supplied in decimal form; i.e., to specify "12%", use 0.12, not 12.

It should go without saying, but make sure that every asset is represented in the three inputs: names of exp_rtn, exp_vol, and the row & colnames of exp_cor.

This function works by finding the Sharpe-optimum equal-weighted portfolio that can be created using the assets passed in. Using that portfolio as a starting point, the function finds the assets A and B such that reallocating ("selling") a small amount (step) of A into asset B ("buying") results in a portfolio whose Sharpe is greater than all other possible assets A and B. The portfolio variable is updated.

This process is repeated with smaller and smaller step sizes. When it is not possible to reallocate step amount of any asset A into any other asset B so as to create a portfolio having a better Sharpe, the Market Portfolio has been reached and the function returns the value.

Examples

# Use the sample dataset "yahoo_adj_prices", provided with the package, for 
# this example. Start with the assumption that the returns that we expect for 
# the next period (day) are the means of the historical log returns observed 
# for a given date window specified by "start_dt" and "end_dt"
end_dt   <- as.Date(zoo::index(xts::last(yahoo_adj_prices))) # latest date
start_dt <- end_dt - 365*3                                   # 3-yr window
end_dt

#> [1] "2020-06-23"

start_dt

#> [1] "2017-06-24"

# From adjusted prices, get the observed daily historical log return
historical_rtn <- yahoo_adj_prices[paste0(start_dt - 1, "/", end_dt)] %>% {
  log(.[-1,] / lag(., k =  1, na.pad = FALSE)[-nrow(.),])
}

# Assume that the returns we expect for the next period (day) are the
# averages of the returns observed during the date window:
exp_rtn <- colMeans(historical_rtn)

# Assume that the volatilities we expect for the next period (day) are the
# standard deviations of the returns observed during the date window:
exp_vol <- dplyr::summarize(
  tibble::as_tibble(historical_rtn),
  dplyr::across(dplyr::everything(), sd, na.rm = TRUE)
) %>%
  purrr::as_vector()

# Assume that the correlations of returns of each asset pair that we expect
# for the next perid (day) are the observed historical correlations:
exp_cor <- stats::cor(historical_rtn, use = "pairwise.complete.obs")

# Calculate the market portfolio:
calculate_market_portfolio(exp_rtn, exp_vol, exp_cor)

#> $sharpe
#> [1] 0.0741225
#> 
#> $weights
#> AAPL  XOM  LNC  MRK  WMT  HOG  RMD AMZN  CMG   FB  IVV 
#> 0.14 0.00 0.00 0.00 0.18 0.00 0.14 0.38 0.16 0.00 0.00 
#> 
#> $ex_return
#> [1] 0.001178704
#> 
#> $ex_volatility
#> [1] 0.01479314
#> 

### Calculate the market portfolio allowing both long & short positions:
calculate_market_portfolio(exp_rtn, exp_vol, exp_cor, allow_shorts = TRUE)

#> $sharpe
#> [1] 0.1185601
#> 
#> $weights
#>       AAPL        XOM        LNC        MRK        WMT        HOG        RMD 
#> 0.22000000 0.00000000 0.00000000 0.00000000 0.08666667 0.00000000 0.10333333 
#>       AMZN        CMG         FB        IVV     s_AAPL      s_XOM      s_LNC 
#> 0.11000000 0.10666667 0.00000000 0.00000000 0.00000000 0.17666667 0.05333333 
#>      s_MRK      s_WMT      s_HOG      s_RMD     s_AMZN      s_CMG       s_FB 
#> 0.00000000 0.00000000 0.09000000 0.00000000 0.00000000 0.00000000 0.05333333 
#>      s_IVV 
#> 0.00000000 
#> 
#> $ex_return
#> [1] 0.0009240314
#> 
#> $ex_volatility
#> [1] 0.007100461
#>