Stock Data

Contains daily high, low, open, close, and volume histories, plus dividend and split events, for 10 (somewhat randomly) selected stocks collected from public sources.

stock_data

Format

A list in which each element is another list, and is named using the identifier of the asset to which its constituent data corresponds (e.g., the element named "AAPL" contains data for Apple, Inc.).

Each of the constituent lists elements contains the following elements:

dividends:

An xts object having a single column -- DividendAmount -- giving the total value of any and all dividends that went ex-div on the date given by the index of the xts.

OHLCV:

A 5-column xts object having column names High, Low, Open, Close, Volume. Each element is the daily historical price (or volume) recorded for the day specified by the index.

splits:

A 2-column xts object having column names Denominator, and Numerator, where Denominator refers to equivalent stock AFTER split, and Numerator refers to stock BEFORE split; e.g, a "2-for-1" split would show Denominator = 2, Numerator = 1. The index of the xts object corresponds to the ex-date of the split described by Denominator and Numerator.

Disclaimer

Absolutely no promise, guarantee, or representation is made by the authors of FinancieR concerning the accuracy or reliability of stock_data, or any other data in the FinancieR package.

Examples

  ## Daily low prices for Chubb from 01 June 2014 to the end of 2015
  chubb_low <- stock_data$CB$prices$Low["2014-06-01/2015"]
  tail(chubb_low)
#>                 Low
#> 2015-12-23 117.2400
#> 2015-12-24 117.6300
#> 2015-12-28 117.4600
#> 2015-12-29 117.7700
#> 2015-12-30 118.0700
#> 2015-12-31 116.7326

  ## All prices data for Pepsico for 2012
  pep_2012 <- stock_data$PEP$prices["2012"]
  tail(pep_2012)
#>             Open  High    Low Close    TWAP    VWAP  Volume PreviousCloseDate
#> 2012-12-21 69.98 70.09 69.234 69.63 69.6429 69.7605 8751056             15694
#> 2012-12-24 69.33 69.60 69.010 69.43 69.2595 69.2725 2071245             15695
#> 2012-12-26 69.41 69.43 68.580 68.84 68.8609 68.8640 3731741             15698
#> 2012-12-27 68.75 68.93 68.185 68.67 68.5106 68.5265 4748031             15700
#> 2012-12-28 68.34 68.69 68.010 68.02 68.3277 68.2912 3764538             15701
#> 2012-12-31 67.78 68.45 67.390 68.43 67.8504 67.9085 5899730             15702
#>            PreviousClose
#> 2012-12-21         70.11
#> 2012-12-24         69.63
#> 2012-12-26         69.43
#> 2012-12-27         68.84
#> 2012-12-28         68.67
#> 2012-12-31         68.02
 
  ## High & Low prices for GD for March 2017
  gd_high_and_low <- stock_data$GD$prices["2017-03", c("High", "Low")]
  tail(gd_high_and_low)
#>               High      Low
#> 2017-03-24 189.895 186.3500
#> 2017-03-27 186.180 183.7201
#> 2017-03-28 189.450 184.8800
#> 2017-03-29 188.920 187.2000
#> 2017-03-30 189.300 188.2000
#> 2017-03-31 188.560 186.4200
 
  ## Dividends paid by Johnson & Johnson in 2017
  stock_data$JNJ$dividends["2017"]
#> NULL
 
  ## Stock splits for AAPL
  stock_data$AAPL$splits
#>            Denominator Numerator
#> 2014-06-09           7         1
#> 2020-08-31           4         1